refactor: move auth.guard and jwt.service to common services folder; create user param decorator
