+++ src/tasks/tasks.module.ts
@@ -1,9 +1,9 @@
 import { Module } from '@nestjs/common';
 import { TasksService } from './tasks.service';
 import { TasksController } from './tasks.controller';
-import { JwtService } from 'src/users/jwtServices/jwt.service';
 import { MongooseModule } from '@nestjs/mongoose';
 import { Task, TaskSchema } from './schemas/task.schema';
+import { JwtService } from 'src/services/jwt.service';
 
 @Module({
   imports: [MongooseModule.forFeature([{ name: Task.name, schema: TaskSchema }])],
