+++ src/users/schemas/user.schema.ts
@@ -3,7 +3,7 @@ import { HydratedDocument } from "mongoose";
 
 export type UserDocument = HydratedDocument<User>
 
-@Schema()
+@Schema({ timestamps: true })
 export class User {
   @Prop({ required: true })
   username: string
@@ -13,6 +13,12 @@ export class User {
 
   @Prop({ required: true })
   password: string
+
+  @Prop()
+  createdAt: Date
+
+  @Prop()
+  updatedAt: Date
 }
 
 export const UserSchema = SchemaFactory.createForClass(User)
