+++ src/users/users.module.ts
@@ -3,12 +3,12 @@ import { UsersService } from './users.service';
 import { UsersController } from './users.controller';
 import { MongooseModule } from '@nestjs/mongoose';
 import { User, UserSchema } from './schemas/user.schema';
-import { JwtService } from './jwtServices/jwt.service';
-import { AuthGuard } from './auth.guard';
+import { AuthGuard } from '../services/auth.guard';
+import { JwtService } from 'src/services/jwt.service';
 
 @Module({
   imports: [MongooseModule.forFeature([{ name: User.name, schema: UserSchema }])],
   controllers: [UsersController],
-  providers: [UsersService, JwtService, AuthGuard],
+  providers: [UsersService, AuthGuard, JwtService],
 })
 export class UsersModule { }
